Switch code to PSR-4
