+++ composer.json
@@ -16,7 +16,9 @@
         "php": ">=5.3.0"
     },
     "autoload": {
-        "psr-0": {"SecureHeaders": ""}
+        "psr-4": {
+            "SecureHeaders\\": "src/"
+        }
     },
     "require-dev": {
         "phpunit/phpunit": "^4.0"
